refactor: ChainGetParentMessages
